pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-allure-step"
version = "0.1.0"
description = "Enhanced logging integration with Allure reports for pytest"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Deekshith Poojary", email = "deekshithpoojary355@gmail.com"}
]
maintainers = [
    {name = "Deekshith Poojary", email = "deekshithpoojary355@gmail.com"}
]
keywords = ["pytest", "allure", "logging", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Pytest",
]
requires-python = ">=3.7"
dependencies = [
    "allure-pytest>=2.9.0",
    "pytest>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "allure-pytest>=2.9.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/deekshith-poojary98/pytest-allure-step"
Repository = "https://github.com/deekshith-poojary98/pytest-allure-step"
Documentation = "https://github.com/deekshith-poojary98/pytest-allure-step#readme"
"Bug Tracker" = "https://github.com/deekshith-poojary98/pytest-allure-step/issues"
Changelog = "https://github.com/deekshith-poojary98/pytest-allure-step/blob/main/CHANGELOG.md"
Source = "https://github.com/deekshith-poojary98/pytest-allure-step"

[tool.setuptools.packages.find]
where = ["."]
include = ["pytest_allure_step*"]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 